+++ src/FluxFlow.Components.Http/HttpComponentTypes.cs
@@ -1,8 +1,0 @@
-using FluxFlow.Engine.Definitions;
-
-namespace FluxFlow.Components.Http;
-
-public static class HttpComponentTypes
-{
-    public static readonly NodeType Client = new("http.client");
-}